separate parsing and building from main script
created build_* modules and docxparser module, to separate the processes
from each other, and eliminate the dependency between them.
MEANING - no more references to html_doc from code that is responsible
for parsing, and no more if-else blocks checking for latex or html build
in the parsing process, etc.
